changes in doc to review

+++ src/main/java/com/psgi/siapweb/domain/service/TypeDocumentService.java
@@ -8,13 +8,13 @@ import com.psgi.siapweb.domain.model.TypeDocument;
 
 public interface TypeDocumentService {
     
-    TypeDocument getTypeDocumentById(Long id_documento);
+    TypeDocument getTypeDocumentById(int id_documento);
     List<TypeDocument> getAllTypeDocuments();
 
     TypeDocument createTypeDocument(TypeDocument typeDocument);
 
-    TypeDocument editTypeDocumentById(Long id_documento, TypeDocument typeDocument);
+    TypeDocument editTypeDocumentById(int id_documento, TypeDocument typeDocument);
 
-    void deleteTypeDocument(Long id_documento);
+    void deleteTypeDocument(int id_documento);
     //ResponseEntity<?> deleteTypeDocument(Long departmentId);
 }
